♻️ refactor(api): move delete_entity/delete_relation to graph router
These two endpoints operate on knowledge-graph entities and relations, so
they belong under the graph API alongside /graph/entity/edit,
/graph/relation/create, etc. They were previously defined in
document_routes.py and mounted under the /documents prefix.

BREAKING CHANGE:
- DELETE /documents/delete_entity   -> DELETE /graph/entity/delete
- DELETE /documents/delete_relation -> DELETE /graph/relation/delete

The frontend does not use these endpoints, so no backward-compat shim is
provided. Request bodies and response models are unchanged.
