+++ lightrag/api/routers/document_routes.py
@@ -25,7 +25,7 @@ from fastapi import (
 from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
 
 from lightrag import LightRAG
-from lightrag.base import DeletionResult, DocProcessingStatus, DocStatus
+from lightrag.base import DocProcessingStatus, DocStatus
 from lightrag.constants import (
     FULL_DOCS_FORMAT_PENDING_PARSE,
     PARSER_ENGINE_LEGACY,
@@ -619,29 +619,6 @@ class DeleteDocRequest(BaseModel):
             raise ValueError("Document IDs must be unique")
 
         return validated_ids
-
-
-class DeleteEntityRequest(BaseModel):
-    entity_name: str = Field(..., description="The name of the entity to delete.")
-
-    @field_validator("entity_name", mode="after")
-    @classmethod
-    def validate_entity_name(cls, entity_name: str) -> str:
-        if not entity_name or not entity_name.strip():
-            raise ValueError("Entity name cannot be empty")
-        return entity_name.strip()
-
-
-class DeleteRelationRequest(BaseModel):
-    source_entity: str = Field(..., description="The name of the source entity.")
-    target_entity: str = Field(..., description="The name of the target entity.")
-
-    @field_validator("source_entity", "target_entity", mode="after")
-    @classmethod
-    def validate_entity_names(cls, entity_name: str) -> str:
-        if not entity_name or not entity_name.strip():
-            raise ValueError("Entity name cannot be empty")
-        return entity_name.strip()
 
 
 class DocStatusResponse(BaseModel):
@@ -4104,81 +4081,6 @@ def create_document_routes(
             logger.error(f"Error clearing cache: {str(e)}")
             logger.error(traceback.format_exc())
             raise HTTPException(status_code=500, detail=str(e))
-
-    @router.delete(
-        "/delete_entity",
-        response_model=DeletionResult,
-        dependencies=[Depends(combined_auth)],
-    )
-    async def delete_entity(request: DeleteEntityRequest):
-        """
-        Delete an entity and all its relationships from the knowledge graph.
-
-        Args:
-            request (DeleteEntityRequest): The request body containing the entity name.
-
-        Returns:
-            DeletionResult: An object containing the outcome of the deletion process.
-
-        Raises:
-            HTTPException: If the entity is not found (404) or an error occurs (500).
-        """
-        try:
-            await check_pipeline_busy_or_raise(rag)
-            result = await rag.adelete_by_entity(entity_name=request.entity_name)
-            if result.status == "not_found":
-                raise HTTPException(status_code=404, detail=result.message)
-            if result.status == "fail":
-                raise HTTPException(status_code=500, detail=result.message)
-            # Set doc_id to empty string since this is an entity operation, not document
-            result.doc_id = ""
-            return result
-        except HTTPException:
-            raise
-        except Exception as e:
-            error_msg = f"Error deleting entity '{request.entity_name}': {str(e)}"
-            logger.error(error_msg)
-            logger.error(traceback.format_exc())
-            raise HTTPException(status_code=500, detail=error_msg)
-
-    @router.delete(
-        "/delete_relation",
-        response_model=DeletionResult,
-        dependencies=[Depends(combined_auth)],
-    )
-    async def delete_relation(request: DeleteRelationRequest):
-        """
-        Delete a relationship between two entities from the knowledge graph.
-
-        Args:
-            request (DeleteRelationRequest): The request body containing the source and target entity names.
-
-        Returns:
-            DeletionResult: An object containing the outcome of the deletion process.
-
-        Raises:
-            HTTPException: If the relation is not found (404) or an error occurs (500).
-        """
-        try:
-            await check_pipeline_busy_or_raise(rag)
-            result = await rag.adelete_by_relation(
-                source_entity=request.source_entity,
-                target_entity=request.target_entity,
-            )
-            if result.status == "not_found":
-                raise HTTPException(status_code=404, detail=result.message)
-            if result.status == "fail":
-                raise HTTPException(status_code=500, detail=result.message)
-            # Set doc_id to empty string since this is a relation operation, not document
-            result.doc_id = ""
-            return result
-        except HTTPException:
-            raise
-        except Exception as e:
-            error_msg = f"Error deleting relation from '{request.source_entity}' to '{request.target_entity}': {str(e)}"
-            logger.error(error_msg)
-            logger.error(traceback.format_exc())
-            raise HTTPException(status_code=500, detail=error_msg)
 
     @router.get(
         "/track_status/{track_id}",
